+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added `--output` support for writing generated changelogs to Markdown files
+
 ## 0.1.0
 
 - Initial changelog generation CLI